backend/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
pydantic 
opencc 
opencc-python-reimplemented 